Version Vieja/Backend_CHPC/requirements.txt
alembic==1.14.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
packaging==24.2
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.2
python-dotenv==0.21.0==1.0.1